DQN scratch implementation for a custom robotic task
The task is to make an e-puck robot learn how to avoid obstacles in a Webots environment. The DQN code implemented here is simple and is written from scratch. Nothing fancy. Just easy to follow. And to top it all, this was implemented on a custom environment ( might not be gym-compliant but I do have plans to make it so! It's a fun project!) The Reward function is designed as per my specifications and can be scaled up if needed. You can play with the hyperparameters to see how it affects the overall performance